app/requirements.txt
django
djangorestframework
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
pillow
gunicorn==19.9.0
django-environ
psycopg2
django-cors-headers
django-model-utils
django-celery
django-celery-email
celery==5.0.5
